refactor: 메서드명 수정, Validate 책임 분리
