+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -3,14 +3,13 @@ package kitchenpos.menu.application;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 import kitchenpos.menu.application.dto.MenuCreateRequest;
 import kitchenpos.menu.application.dto.MenuResponse;
+import kitchenpos.menu.application.dto.MenuValidator;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.repository.MenuProductRepository;
 import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.util.BigDecimalUtil;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,16 +19,16 @@ public class MenuService {
     private final MenuRepository menuRepository;
     private final MenuProductRepository menuProductRepository;
     private final MenuGroupValidator menuGroupValidator;
-    private final ProductTotalPriceValidator productTotalPriceValidator;
+    private final MenuValidator menuValidator;
 
     public MenuService(final MenuRepository menuRepository,
                        final MenuProductRepository menuProductRepository,
                        final MenuGroupValidator menuGroupValidator,
-                       final ProductTotalPriceValidator productTotalPriceValidator) {
+                       final MenuValidator menuValidator) {
         this.menuRepository = menuRepository;
         this.menuProductRepository = menuProductRepository;
         this.menuGroupValidator = menuGroupValidator;
-        this.productTotalPriceValidator = productTotalPriceValidator;
+        this.menuValidator = menuValidator;
     }
 
     @Transactional
@@ -45,42 +44,14 @@ public class MenuService {
         final String name = request.getName();
         final BigDecimal price = request.getPrice();
         final Long menuGroupId = request.getMenuGroupId();
-        final List<MenuProduct> menuProducts = request.getMenuProducts().stream()
-                .map(productRequest -> new MenuProduct(productRequest.getProductId(), productRequest.getQuantity()))
-                .collect(Collectors.toList());
+        final List<MenuProduct> menuProducts = MenuProductMapper.mapToList(request.getMenuProducts());
 
-        validateMenuGroupExist(request.getMenuGroupId());
-        validateMenuPriceNotBiggerThanMenuProductsTotalPrice(price, menuProducts);
+        menuGroupValidator.validateMenuGroupExist(request.getMenuGroupId());
+        menuValidator.validateMenuPriceNotBiggerThanMenuProductsTotalPrice(price, menuProducts);
 
         return Menu.builder(name, price, menuGroupId)
                 .menuProducts(menuProducts)
                 .build();
-    }
-
-    private void validateMenuGroupExist(final Long menuGroupId) {
-        menuGroupValidator.validateMenuGroupExist(menuGroupId);
-    }
-
-    private void validateMenuPriceNotBiggerThanMenuProductsTotalPrice(final BigDecimal price,
-                                                                      final List<MenuProduct> menuProducts) {
-        final BigDecimal menuProductsTotalPrice = getMenuProductsTotalPrice(menuProducts);
-        BigDecimalUtil.valueForCompare(price)
-                .throwIfBiggerThan(menuProductsTotalPrice, IllegalArgumentException::new);
-    }
-
-    private BigDecimal getMenuProductsTotalPrice(final List<MenuProduct> menuProducts) {
-        final List<BigDecimal> menuProductTotalPrices = menuProducts.stream()
-                .map(this::getProductTotalPrice)
-                .collect(Collectors.toList());
-
-        return BigDecimalUtil.sum(menuProductTotalPrices);
-    }
-
-    private BigDecimal getProductTotalPrice(final MenuProduct menuProduct) {
-        return productTotalPriceValidator.getTotalPriceThrowIfNotExist(
-                menuProduct.getProductId(),
-                BigDecimal.valueOf(menuProduct.getQuantity())
-        );
     }
 
     private List<MenuProduct> saveAllMenuProducts(final List<MenuProduct> menuProducts, final Long menuId) {
