--- src/main/java/kitchenpos/menu/application/ProductTotalPriceValidator.java
@@ -1,8 +1,0 @@
-package kitchenpos.menu.application;
-
-import java.math.BigDecimal;
-
-public interface ProductTotalPriceValidator {
-
-    BigDecimal getTotalPriceThrowIfNotExist(final Long productId, final BigDecimal quantity);
-}
