+++ src/main/java/kitchenpos/menu/application/dto/MenuValidator.java
@@ -1,0 +1,42 @@
+package kitchenpos.menu.application.dto;
+
+import java.math.BigDecimal;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.product.domain.Product;
+import kitchenpos.product.domain.repository.ProductRepository;
+import kitchenpos.util.BigDecimalUtil;
+import org.springframework.stereotype.Component;
+
+@Component
+public class MenuValidator {
+
+    private final ProductRepository productRepository;
+
+    public MenuValidator(final ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    public void validateMenuPriceNotBiggerThanMenuProductsTotalPrice(final BigDecimal price,
+                                                                     final List<MenuProduct> menuProducts) {
+        final BigDecimal menuProductsTotalPrice = getMenuProductsTotalPrice(menuProducts);
+        BigDecimalUtil.valueForCompare(price)
+                .throwIfBiggerThan(menuProductsTotalPrice, IllegalArgumentException::new);
+    }
+
+    private BigDecimal getMenuProductsTotalPrice(final List<MenuProduct> menuProducts) {
+        final List<BigDecimal> menuProductTotalPrices = menuProducts.stream()
+                .map(this::getProductTotalPrice)
+                .collect(Collectors.toList());
+
+        return BigDecimalUtil.sum(menuProductTotalPrices);
+    }
+
+    private BigDecimal getProductTotalPrice(final MenuProduct menuProduct) {
+        final Product product = productRepository.findById(menuProduct.getProductId())
+                .orElseThrow(IllegalArgumentException::new);
+
+        return product.calculateTotalPrice(BigDecimal.valueOf(menuProduct.getQuantity()));
+    }
+}
